chore: extract unique ID form modifiers as a separate feature
