+++ test/unit/models/classes/Translation/Listener/TestCreatedEventListenerTest.php
@@ -27,8 +27,8 @@ use core_kernel_classes_Resource;
 use oat\generis\model\data\Ontology;
 use oat\tao\model\featureFlag\FeatureFlagCheckerInterface;
 use oat\tao\model\TaoOntology;
-use oat\taoQtiTest\models\Translation\Listener\TestCreatedEventListener;
-use oat\taoQtiTest\models\Translation\Service\QtiIdentifierRetriever;
+use oat\taoQtiTest\models\UniqueId\Listener\TestCreatedEventListener;
+use oat\taoQtiTest\models\UniqueId\Service\QtiIdentifierRetriever;
 use oat\taoTests\models\event\TestCreatedEvent;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
